refactor: extract some flat input functionality into helpers
