+++ src/components/TextInput/Adornment/types.tsx
@@ -1,15 +1,4 @@
-export enum AdornmentType {
-  Icon = 'icon',
-  Affix = 'affix',
-}
-export enum AdornmentSide {
-  Right = 'right',
-  Left = 'left',
-}
-export enum InputMode {
-  Outlined = 'outlined',
-  Flat = 'flat',
-}
+import { AdornmentSide, AdornmentType } from './enums';
 
 export type AdornmentConfig = {
   side: AdornmentSide;
